'till this point, i think it's working.
we set a map 60x60, and set the hole at [60].
when @pos_x is zero, we see it, since we have no offset from the
edge of the map.
when @pos_x is one, it disappears. now, i just have to certify
it is appearing if i move the camera one unit to the left.
and to check that, i have to make the player stand still and
the board move.

im thinking of a "camera do ... end" syntax on game.rb