creates fire tile on make_fire success.
i'll need to change the tileset drawing process. in this example,
we set_by_coords, so, for example, the road disappears when we burn
in the road - because the actual tile has been replaced on the array.
i must ensure that all ground will be drawn just like the default is.
maybe an array of items for any given position?